Cleanup of remaining Symbolic unit tests
